% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCoverSheetCompleteness.R
\name{checkCoverSheetCompleteness}
\alias{checkCoverSheetCompleteness}
\title{Check Cover Sheet Completeness}
\usage{
checkCoverSheetCompleteness(folder, fileHasHeader, de_map, d2_default_session)
}
\arguments{
\item{folder}{}

\item{fileHasHeader}{}

\item{de_map}{}

\item{d2_default_session}{}
}
\value{

}
\description{
Check Cover Sheet Completeness
}
