% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForCEEValidity.R
\name{checkForCEEValidity}
\alias{checkForCEEValidity}
\title{Check for CEE Validity}
\usage{
checkForCEEValidity(folder, fileHasHeader, de_map, bad_data_values)
}
\arguments{
\item{folder}{}

\item{fileHasHeader}{}

\item{de_map}{}

\item{bad_data_values}{}
}
\value{

}
\description{
Check for CEE Validity
}
