% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForWrongAssessmentType.R
\name{checkForWrongAssessmentType}
\alias{checkForWrongAssessmentType}
\title{Check For Wrong Assessment Type}
\usage{
checkForWrongAssessmentType(folder, fileHasHeader, de_map)
}
\arguments{
\item{folder}{}

\item{fileHasHeader}{}

\item{de_map}{}
}
\value{

}
\description{
Check For Wrong Assessment Type
}
