% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkValueTypeCompliance2.R
\name{checkValueTypeCompliance2}
\alias{checkValueTypeCompliance2}
\title{Check Value Type Compliance 2}
\usage{
checkValueTypeCompliance2(d)
}
\arguments{
\item{d}{}
}
\value{

}
\description{
Check Value Type Compliance 2
}
