% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMS4Validation-package.R
\docType{package}
\name{SIMS4Validation-package}
\alias{SIMS4Validation}
\alias{SIMS4Validation-package}
\title{SIMS4Validation: SIMS 4 validation}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Vladimer Shioshvili \email{vshioshvili@gmail.com}

Authors:
\itemize{
  \item Hiwot Chichaybelu \email{hiwot.fisseha@gmail.com}
}

}
\keyword{internal}
