% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIMSValidationScript.R
\name{SIMSValidationScript}
\alias{SIMSValidationScript}
\title{SIMS Validation Script}
\usage{
SIMSValidationScript(
  out_dir,
  filename,
  file_type,
  idScheme,
  dataElementIdScheme,
  orgUnitIdScheme,
  isoPeriod,
  fileHasHeader,
  secrets,
  dataSets
)
}
\arguments{
\item{out_dir}{}

\item{filename}{}

\item{file_type}{}

\item{idScheme}{}

\item{dataElementIdScheme}{}

\item{orgUnitIdScheme}{}

\item{isoPeriod}{}

\item{fileHasHeader}{}

\item{secrets}{}

\item{dataSets}{}
}
\value{

}
\description{
SIMS Validation Script
}
