% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsValidator.R
\name{simsValidator}
\alias{simsValidator}
\title{SIMS Validator}
\usage{
simsValidator(
  folder,
  filename,
  file_type,
  idScheme,
  dataElementIdScheme,
  orgUnitIdScheme,
  isoPeriod,
  fileHasHeader,
  d2_default_session,
  dataSets
)
}
\arguments{
\item{folder}{}

\item{filename}{}

\item{file_type}{}

\item{idScheme}{}

\item{dataElementIdScheme}{}

\item{orgUnitIdScheme}{}

\item{isoPeriod}{}

\item{fileHasHeader}{}

\item{d2_default_session}{}

\item{dataSets}{}
}
\value{

}
\description{
SIMS Validator
}
